+++ test/bones_api_test_adapter.dart
@@ -65,6 +65,8 @@ TestEntityRepositoryProvider createEntityRepositoryProvider(
     );
 
 abstract class DBTestContainer<D> {
+  String get name;
+
   D? get containerHandler;
 
   FutureOr<bool> setupContainerHandler();
@@ -288,6 +290,10 @@ void runAdapterTests(
         expect(user, isNull);
 
         expect((await userAPIRepository.selectAll()), isEmpty);
+
+        expect((await userAPIRepository.existsID(1)), isFalse);
+        expect((await userAPIRepository.existsID(2)), isFalse);
+        expect((await userAPIRepository.existsID(3)), isFalse);
       }
 
       {
@@ -311,7 +317,27 @@ void runAdapterTests(
         var id = await userAPIRepository.store(user);
         expect(id, equals(1));
 
+        expect(user.id, equals(1));
+        expect(address.id, equals(1));
+
         expect((await userAPIRepository.selectAll()).length, equals(1));
+        expect((await addressAPIRepository.selectAll()).length, equals(1));
+
+        expect((await userAPIRepository.existsID(1)), isTrue);
+        expect((await addressAPIRepository.existsID(1)), isTrue);
+
+        expect((await userAPIRepository.selectByID(1))?.email,
+            equals('joe@$testDomain'));
+        expect((await addressAPIRepository.selectByID(1))?.city,
+            equals('New York'));
+
+        expect((await userAPIRepository.selectByAddress(address)).single.email,
+            equals('joe@$testDomain'));
+        expect(
+            (await userAPIRepository.selectByAddressID(address.id!))
+                .single
+                .email,
+            equals('joe@$testDomain'));
       }
 
       var user2CreationTime = DateTime.utc(2021, 9, 21, 22, 11, 12, 0, 0);
@@ -532,7 +558,7 @@ void runAdapterTests(
         print(transaction);
 
         expect(result!.length, equals(3));
-        expect(transaction.length, equals(7));
+        expect(transaction.length, greaterThanOrEqualTo(7));
         expect(transaction.cachedEntitiesLength, equals(8));
       }
 
@@ -571,7 +597,7 @@ void runAdapterTests(
         print(transaction);
 
         expect(result!.length, equals(3));
-        expect(transaction.length, equals(8));
+        expect(transaction.length, greaterThanOrEqualTo(8));
         expect(transaction.cachedEntitiesLength, equals(8));
       }
 
@@ -715,7 +741,7 @@ void runAdapterTests(
         expect(transaction.isOpen, isTrue);
         expect(transaction.isAborted, isFalse);
         expect(transaction.isCommitted, isTrue);
-        expect(transaction.length, equals(7));
+        expect(transaction.length, greaterThanOrEqualTo(7));
         expect(transaction.abortedError, isNull);
       }
 
@@ -749,7 +775,7 @@ void runAdapterTests(
         expect(transaction.isOpen, isTrue);
         expect(transaction.isAborted, isTrue);
         expect(transaction.isCommitted, isFalse);
-        expect(transaction.length, equals(6));
+        expect(transaction.length, greaterThanOrEqualTo(6));
         expect(transaction.abortedError, isNotNull);
         expect(transaction.abortedError?.reason, equals('Test'));
       }
@@ -865,6 +891,54 @@ void runAdapterTests(
         var address2 = await addressAPIRepository.selectByID(address1.id);
         expect(address2!.toJsonEncoded(), equals(address1.toJsonEncoded()));
       }
+    });
+
+    test('populate', () async {
+      if (dbTestContainer.name == 'mysql') {
+        _log.warning('SKIPPING POPULATE TEST FOR MYSQL: MySQL Driver issue');
+        return;
+      }
+
+      var result = await entityRepositoryProvider.storeAllFromJson({
+        'user': [
+          {
+            'id': 1001,
+            'email': 'extra@mail.com',
+            'password': 'abc789',
+            'roles': [],
+            'wakeUpTime': Time(10, 0),
+            'creationTime': DateTime(2022, 10, 1),
+            'address': 11001,
+          },
+          {
+            'id': 1002,
+            'email': 'extra2@mail.com',
+            'password': 'abc7890',
+            'roles': [],
+            'wakeUpTime': Time(11, 0),
+            'creationTime': DateTime(2022, 10, 2),
+            'address': {
+              'id': 11111,
+              'state': 'EX2',
+              'city': 'Extra2',
+              'street': 'Street x2',
+              'number': 2
+            }
+          },
+        ]
+      });
+
+      expect(result.length, equals(1));
+
+      var usersResult = (result['user'] as List).cast<User>();
+
+      expect(usersResult.length, equals(2));
+
+      expect(usersResult[0].address.id, equals(11001));
+      expect(usersResult[0].address.state, equals('EX'));
+
+      expect(usersResult[1].address.id, equals(11111));
+      expect(usersResult[1].address.state, equals('EX2'));
     });
   });
 }